Director of Skunkworks
As a core member of the senior management team, the post holder will work with the Chief Technology Officer and others to help deliver the missions of NHSX, and the long-term plan commitments to help digitise the NHS.
The Director of Skunkworks will create and lead the Skunkworks function of NHSX. The successful candidate will need to be a strategic
leader who is able to engage in rapid innovation. You will be responsible for leading the team to experiment with data science tools, to build prototypes, to work with real users on exciting new projects, to explore policy and regulatory questions with quickly-built proofs of concept, and to test the art of the possible.
The Skunkworks team will include product managers, developers, user researchers and analysts, and will identify opportunities to build
things that will support our progress against our five missions.
We expect many of the team’s projects to be experimental but those that will help our community to learn. We embrace constant evolution and will use our Skunkworks to drive our understanding of what can and cannot work.